Update Dockerfile to .NET 8 and enable HTTPS
The Dockerfile has been updated to use the official .NET 8 runtime and SDK for building and running the application. The base image is now `mcr.microsoft.com/dotnet/aspnet:8.0`, and the build image is `mcr.microsoft.com/dotnet/sdk:8.0`. The Dockerfile now includes steps to restore dependencies, build the project, publish the application, and set up the final runtime environment. Additionally, the application exposes ports 80 and 443, and includes a step to copy a certificate file (`zenra-finance-back.pfx`) into the final image.

The `zenra-finance-back.http` file has been updated to change the protocol of the `zenra_finance_back_HostAddress` from `http` to `https`, reflecting a move to secure communication over HTTPS.